+++ indico/modules/events/registration/client/js/form/fields/registry.js
@@ -47,7 +47,7 @@ import MultiChoiceInput, {
   multiChoiceGetDataForCondition,
 } from './MultiChoiceInput';
 import NumberInput, {NumberSettings, numberSettingsFormValidator} from './NumberInput';
-import PhoneInput from './PhoneInput';
+import PhoneInput, {PhoneSettings} from './PhoneInput';
 import PictureInput, {PictureSettings} from './PictureInput';
 import SingleChoiceInput, {
   SingleChoiceSettings,
@@ -157,6 +157,7 @@ const fieldRegistry = {
   phone: {
     title: Translate.string('Phone'),
     inputComponent: PhoneInput,
+    settingsComponent: PhoneSettings,
     icon: 'phone',
   },
   country: {